requirements.txt
fastapi
uvicorn
python-dotenv
chromadb
sentence-transformers
pypdf
beautifulsoup4
requests
groq
